refactor: transfer constants
