+++ src/cli/constants.mjs
@@ -1,0 +1,7 @@
+const defaultColorText = '\x1b[0m'
+const appealColorText = '\x1b[0m\x1b[32m'
+const NameColorText = '\x1b[1m\x1b[33m'
+const pathColorText = '\x1b[1m\x1b[34m'
+const errorColorText = '\x1b[1m\x1b[31m'
+
+export { defaultColorText, appealColorText, NameColorText, pathColorText, errorColorText }
